fix 'snack' not working on mobile
could not get snack to work properly on mobile when the snack lived in the .map-container.
created a 'banner' and it lives just about the .map-container.
had to do a lot of jiggery-pokery with grid-layouts to get the layout working correctly.
misc:
all elements fade in on load;
page background is purpely;
planet's label text in now uppercase;
locationIsSurface is now a computed property;