allow user to unselect nodes
The user can now click on a node to unselect it.
Also the user can unselect a node by clicking its name in the panel.

The code has been refactored to remove all the manual DOM manipulation code,
which was a hangover from removing the Cytoscape library.

Created isDefined and isUndefined helper methods.